fix(load): use DOUBLE PRECISION for fact value column
PostgreSQL REAL is single-precision (4-byte, ~7 significant digits).
Python floats are double-precision, so values like 34.85 round-trip
through REAL with a ~1e-6 discrepancy — wide enough to exceed the
1e-9 tolerance in _nan_equal and re-classify unchanged rows as
"updated" on every pipeline run. DOUBLE PRECISION round-trips
Python floats exactly and removes the noise.

SQLite uses dynamic type affinity so the SQLite path (including the
test suite) is unaffected.

Existing PostgreSQL deployments need a one-time column-type change
to migrate stored data:

    ALTER TABLE raw.fact_economic_observations
      ALTER COLUMN value TYPE double precision;

Then a full pipeline run overwrites the previously-truncated values.
For dev environments with no production data, TRUNCATE and re-run
is simpler and produces the same end state.

+++ src/load.py
@@ -92,7 +92,7 @@ def ensure_tables_exist(engine) -> None:
                 series_id   TEXT NOT NULL,
                 series_name TEXT NOT NULL,
                 date        TEXT NOT NULL,
-                value       REAL,
+                value       DOUBLE PRECISION,
                 source      TEXT NOT NULL,
                 PRIMARY KEY (series_id, date)
             )
